refactor(list): move the kind chips into the list column
`KindChips` and its `Chip` are list-column controls — they narrow the
list rather than acting on the shell chrome — but sat under `Header/`,
so the list column reached up two levels to import its own filter row.
Pure `git mv` plus the one import; no markup, class or testid changes.
