+++ src/components/Main/Body/ListColumn/KindChips.tsx
